Write particle data to temporary file piecemeal to improve speed
Previously, process.stiltfoot_array read in all of the particle
data from the matching stilt output files, computed their time
of exit from the domain, if any, then saved the result to file.
This caused a slowdown because all of the particle data was
being held in memory at once. Now, instead, each stilt file is
read, the exit time is calculated, and the result is written to
file before moving on to the next.