x trace coords now made from basis regions
We were assuming that the basis region numbers would always start
at 0, but this is only true because we subtract 1 from the basis
functions before saving the output in openghg_inversions.

There isn't really a reason to modify the basis functions before
outputting them, so if we change this in inversions, our assumption
would be wrong.

This change will avoid this potential problem.