Added code to align basis functions to country
If the footprint and flux used in the inversion had differing
lat/lon coordinates, then the basis functions will be aligned to
the footprint (and "xsensitivity"), while the flux that is saved
will not be aligned to the footprint.

This causes a problem in some cases when trying to compute country
totals.

NOTE: we're using "override" because `inv_out.basis` is a sparse matrix
and currently you can't forward fill in more than one dimension when
using sparse matrices in xarray: https://github.com/pydata/xarray/issues/3445